requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.6
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-cuser==2017.3.16
django-modeltranslation
django-recaptcha==2.0.6
django-tinymce==3.2.0
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
Pillow==8.1.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
